pyproject.toml
[project]
name = "pokesys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "ipykernel>=6.29.5",
    "langchain-openai>=0.3.26",
    "langgraph>=0.4.10",
    "langgraph-supervisor>=0.0.27",
    "openai>=1.91.0",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.34.3",
]
